CyclicOperationProfile Test added
Tests mapper and if creation of base Id
